OF-1061 Always send room subject after history
This change means that room subject changes are no longer held within the
history transcript, but instead are always sent in a distinct action afterward.

Although this means that repeated changes will not be replayed in history, this
seems to be normal behaviour on other servers, and simplified the history replay
code considerably.