Do not send last published item if label prevents
The last published item was always sent, and also
sent with the default label for the system, which is
clearly incorrect on both counts.

This labels correctly and ensures the label is checked.

An ideal solution might send the last visible published
item, but this is considered too difficult given the
existing API.