SASL2: Add password reset mechanism
This is driven by a user property containing the
token in the format <expiry as UNIX time>,
followed by a "|", followed by an alphanumeric
random string. The user property name is
"openfire.password-reset-token".

To aid with testing, the expiry can be "0",
meaning it never expires.

The mechanism, PASSWORD-RESET-TOKEN, is a
client-speaks-first mechanism. The client sends
its requested username, followed by a NUL byte,
followed by the alphanumeric random string
stored on the server.

If this token matches (and has yet to expire),
the "openfire.password.reset" user property
is set with a timestamp of when the reset token
was sent, for debugging purposes.

The PASSWORD-RESET task will then be required.